LteHarqBufferRxD2D: Only retrieve transmitting MAC if necessary
Previously, this module could cause a segfault. This was caused by a vehicle
transmitting a message, immediately before it is removed from the simulation.
Then when a receiver is trying to retrieve the MAC module for the transmitter a
null values returned, causing a segfault. However, this transmitting MAC module
is only used to emit end-to-end delay statistics in non-D2D transmissions.
Thus, for the D2D, multicast case it can be ignored.

Now, the transmitting MAC is only retrieved if it will be necessary, preventing
segfaults in this scenario.